LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_popups_escaping-3-expected.txt

CONSOLE MESSAGE: line 15: Unsafe JavaScript attempt to initiate navigation for frame with URL 'about:blank' from frame with URL 'http://localhost:8800/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_popups_helper-3.html'. The frame attempting navigation is sandboxed, and is therefore disallowed from navigating its ancestors.

CONSOLE MESSAGE: line 15: SecurityError (DOM Exception 18): The operation is insecure.


Harness Error (TIMEOUT), message = null

TIMEOUT Check that popups from a sandboxed iframe escape the sandbox if
       allow-popups-to-escape-sandbox is used Test timed out
